REFACTOR: All component files renamede to index.ts
