+++ components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/VideoFilesArea/index.tsx
