+++ components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/WebFilesArea/index.tsx
