--- components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/folders.tsx
@@ -1,64 +1,0 @@
-import VideoFiles from "@/components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/VideoFilesArea/VideoFiles/videoFiles"
-import VideoFilesArea from "@/components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/VideoFilesArea/videoFilesArea"
-import WebFiles from "@/components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/WebFilesArea/WebFiles/webFiles"
-import WebFilesArea from "@/components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/Folders/WebFilesArea/webFilesArea"
-import Folder from "@/components/Folder/folder"
-
-type foldersType = {
-    topicItems: {
-        subtopic: string,
-        videoList: {
-            title: string,
-            author: string,
-            duration: string,
-            externalPlay: boolean,
-            urlType: "video" | "playlist",
-            urlParam: string
-        }[],
-        readingsList: {
-            title: string,
-            author: string,
-            url: string
-        }[]
-    }[]
-    lastSection: boolean
-}
-
-export default function Folders(props: foldersType) {
-
-    return <>
-        {props.topicItems.map((items, index) => {
-
-            const extraCSS_CloseFolder = (props.lastSection && (props.topicItems.length === index + 1)) ? `
-                ACCESSIBILITY_lastWindowElement
-            ` : ""
-
-            return (
-                <Folder
-                    folderName={items.subtopic}
-                    extraCSS_Open={""}
-                    extraCSS_Close={extraCSS_CloseFolder}
-                    key={index}
-                >
-                    {items.videoList.length > 0 && (
-                        <VideoFilesArea>
-                            <VideoFiles 
-                                videoList={items.videoList}
-                                onlyVideos={items.readingsList.length <= 0 ? true : false}
-                                isLastSection={props.lastSection}
-                            />
-                        </VideoFilesArea>
-                    )}
-                    {items.readingsList.length > 0 && (
-                        <WebFilesArea>
-                            <WebFiles
-                                readingsList={items.readingsList}
-                                isLastSection={props.lastSection}
-                            />
-                        </WebFilesArea>
-                    )}
-                </Folder>
-            )
-        })}
-    </>
-}
