+++ components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections/FolderArea/index.tsx
