+++ components/CustomWindows/MySelfTaughtCurriculum/SectionArea/index.tsx
