+++ components/CustomWindows/MySelfTaughtCurriculum/index.tsx
@@ -1,0 +1,18 @@
+import Sections from "@/components/CustomWindows/MySelfTaughtCurriculum/SectionArea/Sections";
+import WindowBox from "@/components/WindowBox";
+import SectionsArea from "./SectionArea";
+
+export default function MySelfTaughtCurriculumWindow(): JSX.Element {
+    
+    return (
+        <WindowBox
+            windowName="My Self-Taught Curriculum"
+            startVisible={false}
+            CSS_PositionUtilityClass="CSS_MySelfTaughtCurriculum_Position"
+        >
+            <SectionsArea>
+                <Sections />
+            </SectionsArea>
+        </WindowBox>
+    )
+}
