+++ components/WindowBox/CustomWindows/myProjects.tsx
@@ -1,0 +1,13 @@
+import WindowBox from "@/components/WindowBox";
+
+export default function MyProjectsWindow(): JSX.Element {
+    return (
+        <WindowBox
+            windowName="My Projects"
+            startVisible={false}
+            CSS_PositionUtilityClass="CSS_MyProjects_Position"
+        >
+            My Projects
+        </WindowBox>
+    )
+}
